# Perf/improve jobdao query