@algorithmy1 2023-09-07
Thank you for the compliment 🥰

By the way we can improve the query without using `CET`s but we will need to use `limit` 2 times (filter on the page two times).